Added will_paginate, fix routing, bootstrap, +more
Added a whole bunch of stuff. Fixed a whole bunch of stuff.
Was in a seriously mad coding frenzy for the past week and haven't commited a single
chunk. This was definitely the reason for such a large commit.

Going to fully test out rake tasks soon. Most should be completed
Still need to populate the database in order though. And apply some css. Right now it
looks almost worse than it was when I started development.

Also need to develop some tests to run before switching from development mode to
production mode.

Spent a couple days frustrated with the devise gem for not allowing me to remember ldap
authentication with the :rememberable flag. Still need to figure that one out, but lost my
patience with it for the next forseeable year.

Signed-off-by: "Jake Champlin" <jchamplin@thedatacave.com>